32 byte aligned memory and bitpacked dropout mask, might've had an impact, who knows, hard to benchmark at that point, starting to work on convolutional layers, probably about time i abstracted a lot of these matrices into tensors, dimensionality of input is starting to play a big role

+++ LossMetric/Loss.cpp
@@ -6,7 +6,7 @@ void LossMetric::MaeLoss(const float* __restrict x, const float* __restrict y, f
     const __m256 _none = _mm256_set1_ps(-1.0f);
 
     #pragma omp parallel for
-    for (size_t i = 0; i <= rows*cols-8; i += 8) {
+    for (ssize_t i = 0; i <= (ssize_t)(rows*cols)-8; i += 8) {
         const __m256 _x = _mm256_loadu_ps(&x[i]);
         const __m256 _y = _mm256_loadu_ps(&y[i]);
 
@@ -26,7 +26,7 @@ void LossMetric::MseLoss(const float* __restrict x, const float* __restrict y, f
     const __m256 _two = _mm256_set1_ps(2.0f);
 
     #pragma omp parallel for
-    for (size_t i = 0; i <= rows*cols-8; i += 8) {
+    for (ssize_t i = 0; i <= (ssize_t)(rows*cols)-8; i += 8) {
         const __m256 _x = _mm256_loadu_ps(&x[i]);
         const __m256 _y = _mm256_loadu_ps(&y[i]);
